requirements.txt
requests
boto3
pytz
Pillow
youtube-dl
google-api-python-client
py-cord
py-cord[voice]
matplotlib
